Save another byte via a relative jump
We can't change `jmp monitor` to `jr monitor` in input_handler
because it is too far away.  Relative jumps only allow +/-128
bytes.  But we can add a relative jump to an instruction
which also does a relative jump ..

So this:

        ..
        ..
        jr monitor   ; jump relative, two-byte instruction
        ..
        ..
        ..
        jmp monitor  ; absolute jump, three-byte instruction

Becomes:

        ..
        ..
  jmp_monitor:
        jr monitor
        ..
        ..
        ..
        jr jmp_monitor

We execute an extra instruction, but save a single byte.  Worth it.